move: move snowflake to core
