+++ src/main/java/com/fullcar/member/domain/MemberIdService.java
@@ -1,5 +1,6 @@
 package com.fullcar.member.domain;
 
+import com.fullcar.core.id.SnowFlake;
 import org.springframework.stereotype.Service;
 
 @Service
